audio_api_project/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
joblib==1.3.2
mysqlclient==2.2.0
networkx==3.1
nltk==3.8.1
PyAudio==0.2.13
pytz==2023.3
regex==2023.8.8
requests==2.31.0
SpeechRecognition==3.10.0
sqlparse==0.4.4
textrank==0.1.0
tqdm==4.66.1
typing-extensions==4.7.1
urllib3==2.0.4
